操作模块 添加用户发文,添加积分功能

+++ myBBS/user/src/main/java/com/biyou/controller/ScoreController.java
@@ -1,9 +1,15 @@
 package com.biyou.controller;
 
+import com.biyou.pojo.User;
 import com.biyou.pojo.UserScore;
 import com.biyou.service.UserScoreService;
+import com.biyou.utils_entry.Result;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.Cookie;
+import javax.servlet.http.HttpServletRequest;
 
 @CrossOrigin
 @RequestMapping("/score")
@@ -13,11 +19,9 @@ public class ScoreController {
     @Autowired
     private UserScoreService userScoreService;
 
+
     @GetMapping("/increase/{uid}/{inc}")
-    public void increase(@PathVariable("uid") Integer uid, @PathVariable("inc") Integer inc) {
-
+    public void increase(@PathVariable("uid") Long uid, @PathVariable("inc") Integer inc) {
         userScoreService.increase(uid, inc);
-
-
     }
 }
