+++ myBBS/user/src/main/java/com/biyou/dao/UserScoreMapper.java
@@ -13,5 +13,5 @@ public interface UserScoreMapper extends Mapper<UserScore> {
 
     //增加某 用户的 积分+多少
     @Update("update t_userscore set score=score+ #{inc} where id = #{uid}")
-    void increase(@Param("uid") Integer uid,@Param("inc") Integer inc);
+    void increase(@Param("uid") Long uid,@Param("inc") Integer inc);
 }
