+++ myBBS/user/src/main/java/com/biyou/service/UserScoreService.java
@@ -2,5 +2,5 @@ package com.biyou.service;
 
 public interface UserScoreService {
     //对某个用户,增加他的积分
-    void increase (Integer uid,Integer inc);
+    void increase (Long uid,Integer inc);
 }
